Back button on Android could close the wrong popup
Before, the code was checking if there was a popup opened, and then was checking if a popup had a view inside.
But it was not checking if the view was inside the frontest popup ; so it sometimes routes back a "behind" popup.
Additionally, the closed popup / the back routed view was the most behind and not the most in front. 
With this code, it always acts on the visible popup.